move lib.rs to build.rs and solve conflicts
solve conflicts by removing build.rs and making metacall-sys crate regular dependency
