--- source/ports/rs_port/build.rs
@@ -1,4 +1,0 @@
-fn main() {
-    // Find MetaCall library
-    metacall_sys::build();
-}
