+++ source/ports/rs_port/sys/Cargo.toml
@@ -6,9 +6,3 @@ keywords = ["ffi", "bindings", "metacall"]
 edition = "2021"
 license = "Apache-2.0"
 description = "Crate for finding metacall library in the system."
-
-[lib]
-crate-type = ["lib"]
-doctest = false
-name = "metacall_sys"
-path = "src/lib.rs"
